+++ src/dune_rules/lib_mode.mli
@@ -32,14 +32,24 @@ end
 val of_cm_kind : Cm_kind.t -> t
 
 module Map : sig
+  type mode := t
+
   type 'a t =
     { ocaml : 'a Ocaml.Mode.Dict.t
     ; melange : 'a
     }
 
+  val equal : ('a -> 'a -> bool) -> 'a t -> 'a t -> bool
+
   val to_dyn : ('a -> Dyn.t) -> 'a t -> Dyn.t
 
+  val get : 'a t -> mode -> 'a
+
   val map : 'a t -> f:('a -> 'b) -> 'b t
+
+  val make_all : 'a -> 'a t
+
+  val make : byte:'a -> native:'a -> melange:'a -> 'a t
 
   module Set : sig
     type nonrec t = bool t