+++ src/dune_rules/module_compilation.ml
@@ -8,11 +8,7 @@ module CC = Compilation_context
    extension is not .ml or when the .ml and .mli are in different directories.
    This flags makes the compiler think there is a .mli file and will the read
    the cmi file rather than create it. *)
-let force_read_cmi ~(cm_kind : Lib_mode.Cm_kind.t) source_file =
-  let args = [ "-intf-suffix"; Path.extension source_file ] in
-  match cm_kind with
-  | Melange Cmj -> "--bs-read-cmi" :: args
-  | Ocaml (Cmo | Cmx | Cmi) | Melange Cmi -> args
+let force_read_cmi source_file = [ "-intf-suffix"; Path.extension source_file ]
 
 (* Build the cm* if the corresponding source is present, in the case of cmi if
    the mli is not present it is added as additional target to the .cmo
@@ -98,12 +94,18 @@ let build_cm cctx ~force_write_cmi ~precompiled_cmi ~cm_kind (m : Module.t)
   in
   let open Memo.O in
   let* compiler =
-    let+ compiler =
-      match mode with
-      | Ocaml mode -> Memo.return @@ Context.compiler ctx mode
-      | Melange -> Melange_binary.melc sctx ~dir
-    in
-    Result.to_option compiler
+    match mode with
+    | Melange ->
+      let loc = CC.loc cctx in
+      let+ melc = Melange_binary.melc sctx ~loc ~dir in
+      Some melc
+    | Ocaml mode ->
+      Memo.return
+        (let compiler = Context.compiler ctx mode in
+         (* TODO one day remove this silly optimization *)
+         match compiler with
+         | Ok _ as s -> Some s
+         | Error _ -> None)
   in
   (let open Option.O in
   let* compiler = compiler in
@@ -116,7 +118,7 @@ let build_cm cctx ~force_write_cmi ~precompiled_cmi ~cm_kind (m : Module.t)
   in
   let open Memo.O in
   let* extra_args, extra_deps, other_targets =
-    if precompiled_cmi then Memo.return (force_read_cmi ~cm_kind src, [], [])
+    if precompiled_cmi then Memo.return (force_read_cmi src, [], [])
     else
       (* If we're compiling an implementation, then the cmi is present *)
       let public_vlib_module = Module.kind m = Impl_vmodule in
@@ -138,7 +140,7 @@ let build_cm cctx ~force_write_cmi ~precompiled_cmi ~cm_kind (m : Module.t)
         | (Ocaml (Cmo | Cmx) | Melange Cmj), _, _ ->
           let cmi_kind = Lib_mode.Cm_kind.cmi cm_kind in
           Memo.return
-            ( force_read_cmi ~cm_kind src
+            ( force_read_cmi src
             , [ Path.build (Obj_dir.Module.cm_file_exn obj_dir m ~kind:cmi_kind)
               ]
             , [] )
@@ -188,15 +190,7 @@ let build_cm cctx ~force_write_cmi ~precompiled_cmi ~cm_kind (m : Module.t)
     else Command.Args.empty
   in
   let flags, sandbox =
-    let flags =
-      Ocaml_flags.get (CC.flags cctx)
-        (match mode with
-        | Ocaml m -> m
-        | Melange ->
-          (* TODO: define Melange default flags somewhere, should melange rules
-             read from [flags] stanza as well? *)
-          Byte)
-    in
+    let flags = Ocaml_flags.get (CC.flags cctx) mode in
     match Module.pp_flags m with
     | None -> (flags, sandbox)
     | Some (pp, sandbox') ->
@@ -243,7 +237,7 @@ let build_cm cctx ~force_write_cmi ~precompiled_cmi ~cm_kind (m : Module.t)
     (let open Action_builder.With_targets.O in
     Action_builder.with_no_targets (Action_builder.paths extra_deps)
     >>> Action_builder.with_no_targets other_cm_files
-    >>> Command.run ~dir:(Path.build ctx.build_dir) (Ok compiler)
+    >>> Command.run ~dir:(Path.build ctx.build_dir) compiler
           [ Command.Args.dyn flags
           ; cmt_args
           ; Command.Args.S obj_dirs
@@ -333,7 +327,7 @@ let ocamlc_i ~deps cctx (m : Module.t) ~output =
           [ Path.build (Obj_dir.Module.cm_file_exn obj_dir m ~kind:(Ocaml Cmi))
           ]))
   in
-  let ocaml_flags = Ocaml_flags.get (CC.flags cctx) Mode.Byte in
+  let ocaml_flags = Ocaml_flags.get (CC.flags cctx) (Ocaml Byte) in
   let modules = Compilation_context.modules cctx in
   Super_context.add_rule sctx ~dir
     (Action_builder.With_targets.add ~file_targets:[ output ]
@@ -398,7 +392,7 @@ module Alias_module = struct
           (Module_name.to_string shadowed));
     Buffer.contents b
 
-  let of_modules project modules =
+  let of_modules project modules ~alias_module =
     let main_module = Modules.main_module_name modules |> Option.value_exn in
     let aliases =
       Modules.for_alias modules
@@ -409,9 +403,10 @@ module Alias_module = struct
     let shadowed =
       if Dune_project.dune_version project < (3, 5) then []
       else
-        match Modules.alias_module modules with
+        match Modules.lib_interface modules with
         | None -> []
-        | Some alias_module -> [ Module.name alias_module ]
+        | Some m ->
+          if Module.kind m = Alias then [] else [ Module.name alias_module ]
     in
     { main_module; aliases; shadowed }
 end
@@ -420,7 +415,7 @@ let build_alias_module cctx alias_module =
   let modules = Compilation_context.modules cctx in
   let alias_file () =
     let project = Compilation_context.scope cctx |> Scope.project in
-    Alias_module.of_modules project modules |> Alias_module.to_ml
+    Alias_module.of_modules project modules ~alias_module |> Alias_module.to_ml
   in
   let cctx = Compilation_context.for_alias_module cctx alias_module in
   let sctx = Compilation_context.super_context cctx in