+++ test/blackbox-tests/test-cases/merlin/symlinks.t/run.t
@@ -10,18 +10,18 @@ directory in these tests.
 
 Absolute path with symlinks won't match with Dune's root path in which symlinks
 are resolved:
-  $ dune ocaml-merlin --dump-config="$PWD/realsrc" --root="."
+  $ dune ocaml merlin dump-config "$PWD/realsrc" --root="."
   Path $TESTCASE_ROOT/linkroot/realsrc is not in dune workspace ($TESTCASE_ROOT/realroot).
 
 Absolute path with resolved symlinks will match with Dune's root path:
-  $ dune ocaml-merlin \
-  > --dump-config="$(pwd | sed 's/linkroot/realroot/')/realsrc" \
+  $ dune ocaml merlin \
+  > dump-config "$(pwd | sed 's/linkroot/realroot/')/realsrc" \
   > --root="." | head -n 1
   Foo
 
 
 Dune ocaml-merlin also accepts paths relative to the current directory
-  $ dune ocaml-merlin --dump-config="realsrc" --root="." | head -n 1
+  $ dune ocaml merlin dump-config "realsrc" --root="." | head -n 1
   Foo
 
   $ cd realsrc
@@ -29,6 +29,6 @@ Dune ocaml-merlin also accepts paths relative to the current directory
   $ ocamlc_where="$(ocamlc -where)"
   $ export BUILD_PATH_PREFIX_MAP="/OCAMLC_WHERE=$ocamlc_where:$BUILD_PATH_PREFIX_MAP"
 
-  $ dune ocaml-merlin --dump-config="." --root=".." | head -n 2
+  $ dune ocaml merlin dump-config "." --root=".." | head -n 2
   Foo
   ((STDLIB /OCAMLC_WHERE)